Learn Bash Scripting by Building Five Programs
To assist in coding, type in help in the terminal.  Some of the code such as function, if, loop and operators are similar to javascript code. For this lesson, no need login and for each statement entered in the terminal, no need to add a semicolon at the end. Use the word "exit" to exit the program and ls to display the file . After the codes are added in the editor, run the script by entering ./file_name.sh in the terminal.